requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-taggit==3.0.0
importlib-metadata==4.12.0
Markdown==3.4.1
python-decouple==3.6
sqlparse==0.4.2
zipp==3.8.1
